feat. 순위 정렬 및 언어별 상위 5개 기록만 출력
showRecords 실행 시점에 (accuracy + cpm*0.1) 기준으로 정렬, 상위 5개만 출력
